Optimization of AA

+++ include/Pt/Gfx/ImageFormat.h
@@ -122,7 +122,7 @@ class ImageFormat
 
         /** @brief Sets the color in a pixel span with specified alpha values (used by all anti-aliasing raster functions).
         */
-        void copy(Pixel& dst, const Pt::uint8_t* alphas, size_t length,
+        void copy(Pixel& dst, const Pt::uint16_t* alphas, size_t length,
                   const Color& color, CompositionMode mode) const
         { onCopy(dst, alphas, length, color, mode); }
 
@@ -157,7 +157,7 @@ class ImageFormat
         virtual void onCopy(Pixel& dst, const ConstPixel& src, size_t length,
                             CompositionMode mode) const = 0;
 
-        virtual void onCopy(Pixel& dst, const Pt::uint8_t* alphas, size_t length,
+        virtual void onCopy(Pixel& dst, const Pt::uint16_t* alphas, size_t length,
                             const Color& color, CompositionMode mode) const = 0;
 
         virtual void onCopy(ImageView& to, const Point& toPos,
