+++ src/Pt-Gfx/Rgb16Format.cpp
@@ -145,7 +145,7 @@ void Rgb16Format::onCopy(Pixel& to, const ConstPixel& from, size_t length,
     }
 }
 
-void Rgb16Format::onCopy(Pixel& dst, const Pt::uint8_t* alphas, size_t length,
+void Rgb16Format::onCopy(Pixel& dst, const Pt::uint16_t* alphas, size_t length,
                           const Color& color, CompositionMode mode) const
 {
     // ### !!! TODO !!! ###
