+++ src/Pt-Gfx/Yuv12Format.cpp
@@ -122,7 +122,7 @@ void Yuv12Format::onCopy(Pixel& to, const ConstPixel& from, size_t length,
 {
 }
 
-void Yuv12Format::onCopy(Pixel& dst, const Pt::uint8_t* alphas, size_t length,
+void Yuv12Format::onCopy(Pixel& dst, const Pt::uint16_t* alphas, size_t length,
                           const Color& color, CompositionMode mode) const
 {
     // ### !!! TODO !!! ###
